[expo-mail-composer] Move and adapt MailComposerModule from expo to expo-mail-composer
