+++ packages/expo-mail-composer/android/build.gradle
@@ -1,4 +1,3 @@
-
 buildscript {
   repositories {
     google()
@@ -67,4 +66,5 @@ if (new File(rootProject.projectDir.parentFile, 'pubspec.yaml').exists()) {
 
 dependencies {
   expendency "expo-core"
+  api 'com.android.support:appcompat-v7:27.1.1'
 }
